#25 further fixes and a component is removed
