+++ src/Host/WebUI.Test/SourceFormatBuilderService/DeleteDimensionStructureRootAsync_Validation_Should.cs
@@ -25,7 +25,8 @@ namespace WebUI.Test.SourceFormatBuilderService
                .Setup(m => m.GetSourceFormatWithFullDimensionStructureTreeAsync(It.IsAny<SourceFormat>()))
                .ReturnsAsync(_sourceFormatWithoutRootDimensionStructure);
             ISourceFormatBuilderService sourceFormatBuilderService = new SourceFormatBuilderService(
-                _masterDataWebApiClientMock.Object);
+                _masterDataWebApiClientMock.Object,
+                _masterDataValidatorsMock.Object);
             DimensionStructure dimensionStructure = new DimensionStructure
             {
                 Name = "Something root",
@@ -49,7 +50,8 @@ namespace WebUI.Test.SourceFormatBuilderService
         {
             // Arrange
             ISourceFormatBuilderService sourceFormatBuilderService = new SourceFormatBuilderService(
-                _masterDataWebApiClientMock.Object);
+                _masterDataWebApiClientMock.Object,
+                _masterDataValidatorsMock.Object);
             DimensionStructure dimensionStructure = new DimensionStructure
             {
                 Name = "Something root",
