+++ src/Host/WebUI.Test/SourceFormatBuilderService/RemoveItemFromTreeAsync_Validation_Should.cs
@@ -19,7 +19,8 @@ namespace WebUI.Test.SourceFormatBuilderService
         {
             // Arrange
             ISourceFormatBuilderService builderService = new SourceFormatBuilderService(
-                _masterDataWebApiClientMock.Object);
+                _masterDataWebApiClientMock.Object,
+                _masterDataValidatorsMock.Object);
 
             // Act
             Func<Task> action = async () =>
